+++ Suspension/src/Suspension.Core/Models/BasePaths.cs
@@ -26,5 +26,15 @@ namespace MudRunner.Suspension.Core.Models
         /// The base path to response files of dynamic analysis operations.
         /// </summary>
         public static string DynamicAnalysis => Path.Combine(BasePaths.Analysis, "Dynamic");
+
+        /// <summary>
+        /// The base path to response files of dynamic analysis operations that considers half car.
+        /// </summary>
+        public static string HalfCarAnalysis => Path.Combine(BasePaths.DynamicAnalysis, "Half Car");
+
+        /// <summary>
+        /// The base path to response files of dynamic analysis operations that considers half car and six degrees of freedom.
+        /// </summary>
+        public static string HalfCarSixDofAnalysis => Path.Combine(BasePaths.HalfCarAnalysis, "6 Degrees of Freedom");
     }
 }