+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Dynamic/RunDynamicAnalysisRequest.cs
@@ -1,4 +1,5 @@
 ﻿using MudRunner.Commons.DataContracts.Operation;
+using MudRunner.Suspension.DataContracts.Models;
 
 namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
 {
@@ -7,6 +8,11 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
     /// </summary>
     public class RunDynamicAnalysisRequest : OperationRequestBase
     {
+        /// <summary>
+        /// True, if should consider large displacementes. False, otherwise.
+        /// </summary>
+        public bool ConsiderLargeDisplacements { get; set; }
+
         /// <summary>
         /// Unit: s (second).
         /// </summary>
@@ -21,5 +27,10 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
         /// An additional information to be set in the file name that contains the analysis results.
         /// </summary>
         public string AdditionalFileNameInformation { get; set; }
+
+        /// <summary>
+        /// The variables to calculate the base excitation at the system.
+        /// </summary>
+        public BaseExcitation BaseExcitation { get; set; }
     }
 }